PB-32: added wishlist api's custom response

+++ src/wishlists/wishlists.service.ts
@@ -1,5 +1,12 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotAcceptableException,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { CreateApiResponse } from 'src/common/create-response.interface';
+import { ApiGetResponse } from 'src/common/get-response.interface';
 import { Product } from 'src/entity/Product';
 import { User } from 'src/entity/User';
 import { Wishlist } from 'src/entity/Wishlist';
@@ -14,28 +21,78 @@ export class WishlistsService {
     private productsRepository: Repository<Product>,
   ) {}
 
-  async addToWishlist(user: User, productId: string) {
+  async addToWishlist(
+    user: User,
+    productId: string,
+  ): Promise<CreateApiResponse<any>> {
     const product = await this.productsRepository.findOne({
       where: { id: productId },
     });
+    const alreadyAddedProduct = await this.wishlistsRepository.findOne({
+      where: { product: { id: productId } },
+    });
+
+    if (alreadyAddedProduct) {
+      throw new NotAcceptableException(
+        'This product already added to your wishlist',
+      );
+    }
+
     const wishlistItem = this.wishlistsRepository.create({
       product,
       user,
     });
-    return await this.wishlistsRepository.save(wishlistItem);
+    try {
+      const wishlist = await this.wishlistsRepository.save(wishlistItem);
+      if (wishlist) {
+        return {
+          message: 'Added to wishlist successfully',
+          data: [],
+          success: true,
+        };
+      }
+    } catch (error) {
+      console.log(error);
+    }
   }
 
-  async getWishlist(user: User): Promise<Wishlist[]> {
-    return await this.wishlistsRepository.find({
+  async getAllWishlist(user: User): Promise<ApiGetResponse<Wishlist>> {
+    const wishlists = await this.wishlistsRepository.find({
       where: { user: { id: user.id } },
       relations: ['product'],
     });
+    return {
+      success: true,
+      data: wishlists,
+      message: 'Fetched wishlists successfully',
+      meta: {},
+    };
   }
 
-  async removeFromWishlist(id: string): Promise<void> {
+  async removeFromWishlist(
+    id: string,
+    user: User,
+  ): Promise<CreateApiResponse<any>> {
+    const wishlist = await this.wishlistsRepository.findOne({ where: { id } });
+    if (!wishlist) {
+      throw new NotFoundException(`Wishlist with ID ${id} not found`);
+    }
+    const isAuthenticateUser = await this.wishlistsRepository.findOne({
+      where: { user: { id: user.id } },
+    });
+    if (!isAuthenticateUser) {
+      throw new UnauthorizedException(
+        'This user has no access to delete this wishlist',
+      );
+    }
     const result = await this.wishlistsRepository.delete(id);
     if (result.affected === 0) {
       throw new NotFoundException(`Wishlist with ID ${id} not found`);
     }
+    return {
+      success: true,
+      data: [],
+      message: 'Wishlist removed successfully',
+    };
   }
 }
